+++ src-tauri/crates/chatelier/Cargo.toml
@@ -29,6 +29,8 @@ thiserror = "1.0.38"
 serde = { version = "1.0", features = ["derive"], optional = true }
 
 [dev-dependencies]
+eframe = "0.22.0"
+egui = "0.22.0"
 once_cell = "1.17.1"
 reedline-repl-rs = "1.0.3"
 serde_json = "1.0.93"
